Fix solidty-coverage
