+++ .solcover.js
@@ -1,6 +1,10 @@
 module.exports = {
-  port: 8555,
-  testrpcOptions: "-p 8555 -d --accounts 15",
+  providerOptions: {
+    port: 8555,
+    seed: "TestRPC is awesome!",
+    total_accounts: 15,
+    default_balance_ether: 1000000,
+  },
   skipFiles: ["test/"],
   copyPackages: ["openzeppelin-solidity", "zos-lib"],
 };
